+++ src/1_Page/Championship/ui/DashBoard/ui/TeamHistoryTab/ui/LeagueBracket/constant/header.ts
