+++ src/1_Page/Championship/ui/DashBoard/ui/TeamHistoryTab/ui/LeagueBracket/type.d.ts
