+++ src/1_Page/Championship/ui/DashBoard/ui/TeamListTab/type.d.ts
@@ -1,0 +1,3 @@
+type TeamListTabProps = {
+  teamList: ChampionshipTeamInfo[];
+};
